refactor: TC-51 Move folder structure to something sound
