+++ apps/api/src/main/java/dk/treecreate/api/shipmondo/shipment_object_components/Dfm.java
@@ -1,4 +1,4 @@
-package dk.treecreate.api.shipmondo;
+package dk.treecreate.api.shipmondo.shipment_object_components;
 
 public class Dfm {
     private String insurance_type;
